Test Files, Read From a File, and Token Class
Added test files, that should be hand crafted with a test and expected
to check against after, could be easy to automate later. Successfully
read a full file with a buffer size less than and greater than the size
of the file, from the scanner.py. Made a first attempt at a token class.